backend/requirements.txt
# ==============================
# Core ML
# ==============================
torch>=2.0.0
transformers>=4.40.0
sentencepiece>=0.1.99
huggingface_hub>=0.23.0
tokenizers>=0.15.0
safetensors>=0.4.0
pyyaml>=6.0
regex>=2023.0
packaging>=23.0

# ==============================
# Whisper
# ==============================
openai-whisper>=20231117

# ==============================
# FastAPI
# ==============================
fastapi
uvicorn
python-multipart

# ==============================
# LangChain
# ==============================
langchain
langchain-core
langchain-community
langgraph
langsmith
langchain-openai
langchain-mistralai
langchain-ollama
langchain-google-genai
langchain-groq

# ==============================
# Utilities
# ==============================
python-dotenv
pandas
numpy
scipy
tqdm
indic-nlp-library
elevenlabs